Move operator config into pkg/drivers
* Move AWS EBS generator + operator config into pkg/drivers/aws-ebs
* Move common operator hooks to pkg/drivers/common/operator
* Move common generator config to pkg/drivers/common/generator
