+++ pkg/driver/common/operator/hooks.go
@@ -1,4 +1,4 @@
-package defaults
+package operator
 
 import (
 	"fmt"
