+++ pkg/driver/common/operator/replacer.go
@@ -1,4 +1,4 @@
-package defaults
+package operator
 
 import (
 	"os"
