+++ pkg/driver/common/operator/test_manifests/assets.go
