+++ pkg/driver/common/operator/test_manifests/hcp_selector.yaml
